move app to /app
